comment: added a comment

+++ src/main/java/com/medals/medalsbackend/controller/trainer/TrainerController.java
@@ -90,7 +90,8 @@ public class TrainerController {
 
 	@DeleteMapping(value = "/trainer-athlete-connection")
 	public ResponseEntity<Void> removeTrainerAthleteConnection(@RequestParam Long trainerId, @RequestParam Long athleteId) throws Exception {
-		//authorizationService.assertUserHasOwnerAccess(trainerId);
+		// add authorization so both athlete and trainer are able to use this endpoint
+		authorizationService.assertUserHasOwnerAccess(trainerId);
 		athleteService.removeConnection(trainerId, athleteId);
 		return ResponseEntity.status(HttpStatus.ACCEPTED).body(null);
 	}
